Fixed a bug caused by wrong band conditions
A long comment. For banded alignment, it is easy to get Green's formulation
right. We just set scores outside the band to negative infinity. Unfortunately,
it is very tricky to get Suzuki's formulation right. After a long thought, I
believe I have found the right boundary condition for banded alignment.
However, it is difficult to implement in the current code. The fix in this
commit is imperfect. It messes up DP close to the boundaries of the bands.

I might reimplement the extz2_sse and extd2_sse from scratch. It will take
quite a lot of efforts.